requirements.txt
fastapi[standard]
celery[redis]
sqlalchemy[asyncio]
alembic
asyncpg
python-dotenv
pydantic-settings

websockets

# https://www.sphinx-doc.org/en/master/usage/quickstart.html
sphinx

# https://black.readthedocs.io/en/stable/getting_started.html
black

pylint

# https://flake8.pycqa.org/en/latest/index.html#quickstart
flake8

# https://mypy.readthedocs.io/en/stable/getting_started.html
mypy


prometheus-fastapi-instrumentator

passlib
bcrypt==4.0.1
python-jose
motor


pytest
pytest-cov
pytest-mock
pytest-asyncio

httpx

aiokafka